Version 1.0
This is the version that will be presented to the class on 6/19/2020. It cuts some features that were originally in development to save time, but also adds plenty of new features and visual enhancements. There will potentially be another version where experimental features will be fleshed out (tally of constellations seen). Also attached a new background image under the same name.